requirements.txt
# Auto Generated
bs4
gig-nuuuwan
pytest
selenium
tweepy
utils-nuuuwan
